requirements.txt
dj-database-url==0.5.0
Django==2.0.9
gunicorn==19.9.0
pytz==2018.7
whitenoise==4.1.2
psycopg2==2.7.6.1
requests==2.20.1
numpy==1.15.4
pandas==0.23.4
six==1.11.0
SQLAlchemy==1.2.14
APScheduler==3.2.0
django-chartjs
django-debug-toolbar
python-decouple
djangorestframework
python-dotenv